Rename app to competences
