+++ arbre_competences/urls.py
@@ -17,7 +17,7 @@ Including another URLconf
 """
 from django.conf.urls import url, include
 from django.contrib import admin
-from app import urls
+from competences import urls
 
 urlpatterns = [
     url(r'^admin/?', admin.site.urls),
