+++ competences/admin.py
