+++ competences/forms.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+# -*-coding:utf-8 -*-
+from django import forms
