+++ competences/templates/app/affichage_details.html
