+++ competences/templates/app/liste_categories.html
