+++ competences/urls.py
