+++ competences/views.py
